added wasmraypy/__init__.pt file (a replacement for wrapper.py)
