+++ tools/code_generation/filesGeneration.py
@@ -46,11 +46,11 @@ def add_text_to_file(file_path: Path, _string: str) -> None:
 
 def generate_wasm_array_classes_code() -> str:
     _string = ""
-    _string += array_generation.wasm_array_string + '\n\n'
-    _string += array_generation.struct_array_string + '\n\n'
+    _string += array_generation.wasm_array_string + '\n'
+    _string += array_generation.struct_array_string + '\n'
 
     for primitive_array_metadata in array_generation.primitive_array_classes_metadata:
-        _string += array_generation.generate_primitive_array_class(primitive_array_metadata) + '\n\n'
+        _string += array_generation.generate_primitive_array_class(primitive_array_metadata) + '\n'
 
     return _string
 
@@ -61,7 +61,7 @@ def generate_structs_aliases_code(structs_api, aliases_api) -> str:
         if struct_api['name'] in wrapped_structures_names:
             continue
         wrapped_enums_names.append(struct_api['name'])
-        _string += struct_generation.generate_struct_code(struct_api) + '\n'
+        _string += struct_generation.generate_struct_code(struct_api)
 
         struct_aliases = struct_generation.does_struct_name_has_alias(struct_api['name'], aliases_api)
         for alias_api in struct_aliases:
@@ -129,8 +129,17 @@ raygui_api_functions = raygui_api['functions']
 
 # -----------------------------------------
 # generate all the files for wasmraypy
-generate_file(WASMRAYPY_FOLDER_PATH / 'structures/__init__.py')
-add_text_to_file(WASMRAYPY_FOLDER_PATH / 'structures/__init__.py', "from ..utils import struct_clone\n\n")
-add_text_to_file(WASMRAYPY_FOLDER_PATH / 'structures/__init__.py', generate_wasm_array_classes_code())
-add_text_to_file(WASMRAYPY_FOLDER_PATH / 'structures/__init__.py',
+generate_file(WASMRAYPY_FOLDER_PATH / '__init__.py')
+add_text_to_file(WASMRAYPY_FOLDER_PATH / '__init__.py',
+"""def struct_clone(source, a):
+    if not a:
+        a = _mod._malloc(source.size)
+    _mod._memcpy(a, source._address, source.size)
+    out = source.__class__(address=a, to_alloc=False)
+    return out
+
+"""
+                 )
+add_text_to_file(WASMRAYPY_FOLDER_PATH / '__init__.py', generate_wasm_array_classes_code())
+add_text_to_file(WASMRAYPY_FOLDER_PATH / '__init__.py',
                  generate_structs_aliases_code(raylib_api_structs, raylib_api_aliases))
